refactor: split the audit exports page into a form, a list and a requests composable
