+++ eslint.config.js
@@ -256,7 +256,6 @@ export default defineConfigWithVueTs(
             'resources/js/components/UserStatusDialog.vue',
             'resources/js/composables/useAttachmentUploads.test.ts',
             'resources/js/layouts/MainLayout.vue',
-            'resources/js/pages/teams/AuditExports.vue',
             'resources/js/pages/teams/Groups.vue',
         ],
         rules: {
